SQL: set section_id to double in street_references table

+++ sql_scripts/create_street_references_table.sql
@@ -1,6 +1,6 @@
 CREATE table street_references (
     ID INTEGER NOT NULL AUTO_INCREMENT,
     street_id INTEGER,
-    section_id INTEGER,
+    section_id DOUBLE,
     PRIMARY KEY(ID)
 );
